chg: usr: force the dtype at int for array in _init_with_raw

+++ ruleskit/activation.py
@@ -337,7 +337,7 @@ class Activation(ABC):
         >>> Activation._array_to_int(np.array([0, 1, 1, 0]))
         6  # the binary number '0110' is 6 in base 10
         """
-        return int("".join(str(i) for i in value), 2)
+        return int("".join(str(i) for i in value.astype('int')), 2)
 
     @property
     def raw(self) -> np.ndarray:
